requirements.txt
absl-py==1.4.0
aiohttp==3.8.1
aiosignal==1.3.1
alpaca-trade-api==3.0.0
altair==4.2.2
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
beaker-pyteal==0.6.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
bokeh==2.4.3
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorcet==3.0.1
contourpy==1.0.7
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
deprecation==2.1.0
docstring-parser==0.14.1
entrypoints==0.4
fake-useragent==1.1.3
feedparser==6.0.10
flatbuffers==23.3.3
fonttools==4.39.4
frozendict==2.3.7
frozenlist==1.3.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.3
h5py==3.8.0
holoviews==1.16.0
html5lib==1.1
hvplot==0.8.3
idna==3.4
importlib-metadata==6.4.1
Jinja2==3.1.2
jsonschema==4.17.3
keras==2.11.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
msgpack==1.0.3
multidict==6.0.4
multitasking==0.0.11
numpy==1.24.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
panel==0.14.4
param==1.13.0
parse==1.19.0
Pillow==9.5.0
plotly==5.14.1
protobuf==3.19.6
py-algorand-sdk==2.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodomex==3.17
pyct==0.5.0
pydeck==0.8.1b0
pyee==8.2.2
Pygments==2.15.0
Pympler==1.0.1
PyNaCl==1.5.0
pyotp==2.8.0
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyrsistent==0.19.3
pyteal==0.23.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyviz-comms==2.2.1
PyYAML==6.0
requests==2.28.2
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==13.3.4
rsa==4.9
semantic-version==2.10.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
streamlit==1.21.0
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow @ https://storage.googleapis.com/tensorflow/mac/cpu/tensorflow-1.12.0-py3-none-any.whl
tensorflow-estimator==2.11.0
tensorflow-macos==2.11.0
termcolor==2.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
w3lib==2.1.1
watchdog==3.0.0
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
yahoo-fin==0.8.9.1
yarl==1.9.1
yfinance==0.2.18
zipp==3.15.0
